requirements.txt
matplotlib>=3.7.1
pip>=23.0.1
requests>=2.28.2
python-dotenv>=1.0.0